+++ src/api/api.types.ts
@@ -45,6 +45,11 @@ export type CreateCorpusData = {
   corpus_id: string;
 };
 
+/** Data in the create-metadata response */
+export type CreateMetadataData = {
+  resource_id: string;
+};
+
 /** Data in the resource-info response, if no corpus_id param is given */
 export type ResourceInfoAllData = {
   resources: MinkResponse<ResourceInfo>[];
@@ -68,7 +73,7 @@ export type ResourceData = {
   source_files: FileMeta[];
 };
 
-export type ResourceType = "corpus";
+export type ResourceType = "corpus" | "metadata";
 
 /** Job status for a resource */
 // There's more but we're not using everything.