+++ src/library/LibraryView.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import useLocale from "@/i18n/locale.composable";
 import PadButton from "@/components/PadButton.vue";
 import LayoutSection from "@/components/LayoutSection.vue";
 import PendingContent from "@/spin/PendingContent.vue";
@@ -13,10 +14,11 @@ import useSpin from "@/spin/spin.composable";
 import useCreateCorpus from "@/corpus/createCorpus.composable";
 
 const resourceStore = useResourceStore();
-const { requireAuthentication, isAuthenticated } = useAuth();
+const { requireAuthentication, isAuthenticated, canUserAdmin } = useAuth();
 const { loadResources } = useResources();
 const { createFromUpload } = useCreateCorpus();
 const { spin } = useSpin();
+const { th } = useLocale();
 
 requireAuthentication(loadResources);
 
@@ -59,5 +61,25 @@ async function createCorpusFromFiles(files: FileList) {
         :primary="!resourceStore.hasCorpora"
       />
     </PendingContent>
+  </LayoutSection>
+
+  <LayoutSection v-if="canUserAdmin" :title="$t('metadata')">
+    <HelpBox>
+      <p>{{ $t("library.help.metadata") }}</p>
+    </HelpBox>
+
+    <div class="my-4 flex flex-wrap gap-4">
+      <template v-for="(metadata, id) of resourceStore.metadatas" :key="id">
+        <PadButton :to="`/library/metadata/${id}`">
+          <strong>{{ th(metadata.name) || id }}</strong>
+          {{ metadata.publicId }}
+        </PadButton>
+      </template>
+
+      <PadButton to="/library/metadata/new">
+        <icon :icon="['far', 'square-plus']" size="2xl" class="mb-2" />
+        {{ $t("metadata.new") }}
+      </PadButton>
+    </div>
   </LayoutSection>
 </template>